Read filtered environment endpoints for persistence
Instead of reading modules from db to determine sensors and variables,
just read the environment/ID/filtered/VARIABLE endpoint and persist
whatever is published from it.

This allows persistence to be loosely coupled to sensors. Persistence no
longer has to know or care about sensor implementation details, it just cares
about the environmental variables that are published to the filtered endpoints.
This allows environmental variables to be produced from one or more sensor
inputs, or from other nodes, or whatever -- it doesn't matter.